# openapi.yaml
openapi: 3.0.3
info:
  title: RAG Service
  version: 2.1.0
paths:
...
  /ingestion/jobs:
    post:
      tags:
      - ingestion-jobs
      summary: Create Job
      operationId: create_job_ingestion_jobs_post
      parameters:
      - name: run
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Run
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestionJobCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJobInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - ingestion-jobs
      summary: List Jobs
      operationId: list_jobs_ingestion_jobs_get
      parameters:
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: data_wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
      - name: private_db_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Private Db Id
      - name: session_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
      - name: app_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: App Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJobList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ingestion/jobs/{job_id}:
    get:
      tags:
      - ingestion-jobs
      summary: Get Job
      operationId: get_job_ingestion_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: integer
          title: Job Id
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJobInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ingestion/jobs/{job_id}/run:
    post:
      tags:
      - ingestion-jobs
      summary: Run Job
      operationId: run_job_ingestion_jobs__job_id__run_post
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: integer
          title: Job Id
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJobInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ingestion/jobs/{job_id}/runs:
    get:
      tags:
      - ingestion-jobs
      summary: List Runs
      operationId: list_runs_ingestion_jobs__job_id__runs_get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: integer
          title: Job Id
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJobRuns'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ingestion/jobs/presets:
    get:
      tags:
      - ingestion-jobs
      summary: Ingestion Job Presets
      operationId: ingestion_job_presets_ingestion_jobs_presets_get
      parameters:
      - name: wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
      - name: data_wallet_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
      - name: app_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: App Id
      - name: kb_key
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Kb Key
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJobPreset'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IngestionJobCreate:
      properties:
        wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallet Id
          description: 开发者钱包 ID（兼容字段；推荐使用 Authorization）
        data_wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
          description: 数据归属钱包 ID（业务用户）
        private_db_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Private Db Id
          description: 私有库 ID（session_id 二选一）
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
          description: 业务会话 ID（私有库绑定）
        app_id:
          type: string
          title: App Id
          description: Plugin app_id
        kb_key:
          type: string
          title: Kb Key
          description: KB key
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
          description: MinIO URL (minio://bucket/key)
        content:
          anyOf:
          - type: string
          - type: 'null'
          title: Content
          description: Inline text content (will be stored to MinIO)
        filename:
          anyOf:
          - type: string
          - type: 'null'
          title: Filename
          description: Optional filename for inline content
        file_type:
          anyOf:
          - type: string
          - type: 'null'
          title: File Type
          description: Optional file type override
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
          description: Optional metadata
        options:
          additionalProperties: true
          type: object
          title: Options
          description: Job options
      type: object
      required:
      - app_id
      - kb_key
      title: IngestionJobCreate
    IngestionJobInfo:
      properties:
        id:
          type: integer
          title: Id
        wallet_id:
          type: string
          title: Wallet Id
        data_wallet_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Wallet Id
        private_db_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Private Db Id
        app_id:
          type: string
          title: App Id
        kb_key:
          type: string
          title: Kb Key
        job_type:
          type: string
          title: Job Type
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        file_type:
          anyOf:
          - type: string
          - type: 'null'
          title: File Type
        status:
          type: string
          title: Status
        options_json:
          anyOf:
          - type: string
          - type: 'null'
          title: Options Json
        result_json:
          anyOf:
          - type: string
          - type: 'null'
          title: Result Json
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated At
        started_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Started At
        finished_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Finished At
      type: object
      required:
      - id
      - wallet_id
      - app_id
      - kb_key
      - job_type
      - status
      title: IngestionJobInfo
    IngestionJobList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/IngestionJobInfo'
          type: array
          title: Items
      type: object
      title: IngestionJobList
    IngestionJobPreset:
      properties:
        bucket:
          type: string
          title: Bucket
        prefix:
          type: string
          title: Prefix
        recent_keys:
          items:
            type: string
          type: array
          title: Recent Keys
      type: object
      required:
      - bucket
      - prefix
      title: IngestionJobPreset
    IngestionJobRunItem:
      properties:
        id:
          type: integer
          title: Id
        job_id:
          type: integer
          title: Job Id
        status:
          type: string
          title: Status
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
        meta_json:
          anyOf:
          - type: string
          - type: 'null'
          title: Meta Json
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
      type: object
      required:
      - id
      - job_id
      - status
      title: IngestionJobRunItem
    IngestionJobRuns:
      properties:
        items:
          items:
            $ref: '#/components/schemas/IngestionJobRunItem'
          type: array
          title: Items
      type: object
      title: IngestionJobRuns
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...